examples/9. Deployed agent/requirements.txt
agno>=2.0.3
fastapi[standard]
uvicorn
openai
python-dotenv